ai_agents/requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==4.2.20
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.0
pytz==2025.2
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
